+++ crates/assembly-syntax/src/sema/errors.rs
@@ -67,13 +67,19 @@ pub enum SemanticAnalysisError {
         #[label]
         prev_span: SourceSpan,
     },
+    #[error("invalid invocation target")]
+    #[diagnostic(help("path contains an invalid component"))]
+    InvalidInvokePath {
+        #[label]
+        span: SourceSpan,
+    },
     #[error("invalid program: procedure exports are not allowed")]
     #[diagnostic(help("perhaps you meant to use `proc` instead of `export`?"))]
     UnexpectedExport {
         #[label]
         span: SourceSpan,
     },
-    #[error("invalid enum type representation: underlying type must be an integral type")]
+    #[error("invalid enum type representation: underlying type must be an integral or felt type")]
     #[diagnostic()]
     InvalidEnumRepr {
         #[label]